+++ src/main/java/com/joojoo/api/user/application/UserUseCaseService.java
@@ -1,26 +1,28 @@
 package com.joojoo.api.user.application;
 
 import com.joojoo.api.block.application.port.in.DeleteBlockUseCase;
-import com.joojoo.api.block.domain.repository.BlockRepository;
 import com.joojoo.api.jwt.application.JwtTokenUseCase;
 import com.joojoo.api.tradeLog.application.port.in.DeleteTradeLogUseCase;
-import com.joojoo.api.tradeLog.domain.repository.TradeLogRepository;
 import com.joojoo.api.user.application.auth.withdraw.out.SocialUnlink;
 import com.joojoo.api.user.application.port.in.GetUserUseCase;
 import com.joojoo.api.user.application.port.in.LogoutUseCase;
 import com.joojoo.api.user.application.port.in.WithdrawUserUseCase;
+import com.joojoo.api.user.application.port.out.FilePort;
 import com.joojoo.api.user.domain.model.entity.User;
+import com.joojoo.api.user.domain.model.enums.DefaultProfileImage;
 import com.joojoo.api.user.domain.repository.UserRepository;
-import com.joojoo.api.user.domain.service.auth.AppleClientSecret;
-import com.joojoo.api.user.domain.service.auth.KakaoClientSecret;
-import com.joojoo.api.util.random.GeneratorRandom;
+import com.joojoo.api.user.presentation.dto.response.DefaultProfileImageResponse;
+import com.joojoo.api.user.presentation.dto.response.UserInfoResponse;
 import com.joojoo.global.exception.handleException.users.UserNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Arrays;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -32,11 +34,39 @@ public class UserUseCaseService implements GetUserUseCase, WithdrawUserUseCase, 
     private final DeleteBlockUseCase deleteBlockUseCase;
     private final DeleteTradeLogUseCase deleteTradeLogUseCase;
     private final Map<String, SocialUnlink> socialUnlink;
+    private final FilePort filePort;
 
     @Override
     public User getUser(Long userId) {
         return userRepository.findById(userId)
                 .orElseThrow(UserNotFoundException::new);
+    }
+
+    @Override
+    public User getUserReference(Long userId) {
+        try {
+            return userRepository.getReferenceById(userId);
+        } catch (EntityNotFoundException e) {
+            throw new UserNotFoundException();
+        }
+    }
+
+    @Override
+    public UserInfoResponse getUserInfo(Long userId) {
+        User user = this.getUser(userId);
+        return UserInfoResponse.of(user);
+    }
+
+    @Override
+    public DefaultProfileImageResponse getDefaultProfileImages() {
+        return DefaultProfileImageResponse.of(Arrays.stream(DefaultProfileImage.values())
+                .map(img -> new DefaultProfileImageResponse.ProfileImage(
+                        img.name(),
+                        filePort.getFullUrl(img.getFileName()),
+                        img.getDescription()
+                ))
+                .collect(Collectors.toList())
+        );
     }
 
     @Override
