--- src/main/java/com/joojoo/api/user/domain/provider/fileService.java
@@ -1,5 +1,0 @@
-package com.joojoo.api.user.domain.provider;
-
-public interface fileService {
-    String getFullUrl(String fileName);
-}
