+++ src/main/java/com/joojoo/api/user/infrastructure/image/FileStorageAdapter.java
@@ -1,11 +1,11 @@
 package com.joojoo.api.user.infrastructure.image;
 
-import com.joojoo.api.user.domain.provider.fileService;
+import com.joojoo.api.user.application.port.out.FilePort;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class fileServiceImpl implements fileService {
+public class FileStorageAdapter implements FilePort {
 
     @Value("${STORAGE_ENDPOINT}")
     private String endPoint;
